+++ contracts/modules/Experimental/Burn/TrackedRedemptionFactory.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.4.24;
 
 import "./TrackedRedemption.sol";
-import "../ModuleFactory.sol";
+import "../../ModuleFactory.sol";
 
 /**
  * @title Factory for deploying GeneralTransferManager module
@@ -27,13 +27,14 @@ contract TrackedRedemptionFactory is ModuleFactory {
     }
 
     /**
-     * @notice used to launch the Module with the help of factory
-     * @return address Contract address of the Module
+     * @notice Used to launch the Module with the help of factory
+     * @return Address Contract address of the Module
      */
     function deploy(bytes /* _data */) external returns(address) {
         if (setupCost > 0)
-            require(polyToken.transferFrom(msg.sender, owner, setupCost), "Failed transferFrom because of sufficent Allowance is not provided");
+            require(polyToken.transferFrom(msg.sender, owner, setupCost), "Insufficent allowance or balance");
         address trackedRedemption = new TrackedRedemption(msg.sender, address(polyToken));
+        /*solium-disable-next-line security/no-block-members*/
         emit GenerateModuleFromFactory(address(trackedRedemption), getName(), address(this), msg.sender, setupCost, now);
         return address(trackedRedemption);
     }