+++ contracts/modules/Experimental/TransferManager/SingleTradeVolumeRestrictionTMFactory.sol
@@ -1,8 +1,9 @@
 pragma solidity ^0.4.24;
 
-import "./../ModuleFactory.sol";
+import "./../../ModuleFactory.sol";
 import "./SingleTradeVolumeRestrictionTM.sol";
-import "../../libraries/Util.sol";
+import "../../../libraries/Util.sol";
+
 /**
  * @title Factory for deploying SingleTradeVolumeRestrictionManager
  */
@@ -28,7 +29,7 @@ contract SingleTradeVolumeRestrictionTMFactory is ModuleFactory {
     }
 
     /**
-    * @notice used to launch the Module with the help of factory
+    * @notice Used to launch the Module with the help of factory
     * @return address Contract address of the Module
     */
     function deploy(bytes _data) external returns(address) {
@@ -37,7 +38,9 @@ contract SingleTradeVolumeRestrictionTMFactory is ModuleFactory {
         SingleTradeVolumeRestrictionTM singleTradeVolumeRestrictionManager = new SingleTradeVolumeRestrictionTM(msg.sender, address(polyToken));
 
         require(Util.getSig(_data) == singleTradeVolumeRestrictionManager.getInitFunction(), "Provided data is not valid");
+        /*solium-disable-next-line security/no-low-level-calls*/
         require(address(singleTradeVolumeRestrictionManager).call(_data), "Unsuccessful call");
+        /*solium-disable-next-line security/no-block-members*/
         emit GenerateModuleFromFactory(address(singleTradeVolumeRestrictionManager), getName(), address(this), msg.sender, setupCost, now);
         return address(singleTradeVolumeRestrictionManager);
     }
@@ -57,6 +60,7 @@ contract SingleTradeVolumeRestrictionTMFactory is ModuleFactory {
     * @return string
     */
     function getInstructions() external view returns(string) {
+        /*solium-disable-next-line max-len*/
         return "Allows an issuer to impose volume restriction on a single trade. Init function takes two parameters. First parameter is a bool indicating if restriction is in percentage. The second parameter is the value in percentage or amount of tokens";
     }