+++ contracts/modules/ModuleFactory.sol
@@ -31,7 +31,13 @@ contract ModuleFactory is IModuleFactory, Ownable {
     event ChangeFactorySetupFee(uint256 _oldSetupCost, uint256 _newSetupCost, address _moduleFactory);
     event ChangeFactoryUsageFee(uint256 _oldUsageCost, uint256 _newUsageCost, address _moduleFactory);
     event ChangeFactorySubscriptionFee(uint256 _oldSubscriptionCost, uint256 _newMonthlySubscriptionCost, address _moduleFactory);
-    event GenerateModuleFromFactory(address _module, bytes32 indexed _moduleName, address indexed _moduleFactory, address _creator, uint256 _timestamp);
+    event GenerateModuleFromFactory(
+        address _module,
+        bytes32 indexed _moduleName,
+        address indexed _moduleFactory,
+        address _creator,
+        uint256 _timestamp
+    );
     event ChangeSTVersionBound(string _boundType, uint8 _major, uint8 _minor, uint8 _patch);
 
     /**
@@ -46,7 +52,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice used to change the fee of the setup cost
+     * @notice Used to change the fee of the setup cost
      * @param _newSetupCost new setup cost
      */
     function changeFactorySetupFee(uint256 _newSetupCost) public onlyOwner {
@@ -55,7 +61,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice used to change the fee of the usage cost
+     * @notice Used to change the fee of the usage cost
      * @param _newUsageCost new usage cost
      */
     function changeFactoryUsageFee(uint256 _newUsageCost) public onlyOwner {
@@ -64,7 +70,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice used to change the fee of the subscription cost
+     * @notice Used to change the fee of the subscription cost
      * @param _newSubscriptionCost new subscription cost
      */
     function changeFactorySubscriptionFee(uint256 _newSubscriptionCost) public onlyOwner {
@@ -74,7 +80,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice Update the title of the ModuleFactory
+     * @notice Updates the title of the ModuleFactory
      * @param _newTitle New Title that will replace the old one.
      */
     function changeTitle(string _newTitle) public onlyOwner {
@@ -83,7 +89,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice Update the description of the ModuleFactory
+     * @notice Updates the description of the ModuleFactory
      * @param _newDesc New description that will replace the old one.
      */
     function changeDescription(string _newDesc) public onlyOwner {
@@ -92,7 +98,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice Update the name of the ModuleFactory
+     * @notice Updates the name of the ModuleFactory
      * @param _newName New name that will replace the old one.
      */
     function changeName(bytes32 _newName) public onlyOwner {
@@ -101,7 +107,7 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice Update the version of the ModuleFactory
+     * @notice Updates the version of the ModuleFactory
      * @param _newVersion New name that will replace the old one.
      */
     function changeVersion(string _newVersion) public onlyOwner {
@@ -115,8 +121,11 @@ contract ModuleFactory is IModuleFactory, Ownable {
      * @param _newVersion new version array
      */
     function changeSTVersionBounds(string _boundType, uint8[] _newVersion) external onlyOwner {
-        require(keccak256(abi.encodePacked(_boundType)) == keccak256(abi.encodePacked("lowerBound")) || keccak256(abi.encodePacked(_boundType)) == keccak256(abi.encodePacked("upperBound")),
-        "Must be a valid bound type");
+        require(
+            keccak256(abi.encodePacked(_boundType)) == keccak256(abi.encodePacked("lowerBound")) ||
+            keccak256(abi.encodePacked(_boundType)) == keccak256(abi.encodePacked("upperBound")),
+            "Must be a valid bound type"
+        );
         require(_newVersion.length == 3);
         if (compatibleSTVersionRange[_boundType] != uint24(0)) { 
             uint8[] memory _currentVersion = VersionUtils.unpack(compatibleSTVersionRange[_boundType]);
@@ -127,15 +136,15 @@ contract ModuleFactory is IModuleFactory, Ownable {
     }
 
     /**
-     * @notice use to get the lower bound
+     * @notice Used to get the lower bound
      * @return lower bound
      */
     function getLowerSTVersionBounds() external view returns(uint8[]) {
-       return VersionUtils.unpack(compatibleSTVersionRange["lowerBound"]);
+        return VersionUtils.unpack(compatibleSTVersionRange["lowerBound"]);
     }
 
     /**
-     * @notice use to get the upper bound
+     * @notice Used to get the upper bound
      * @return upper bound
      */
     function getUpperSTVersionBounds() external view returns(uint8[]) {