+++ contracts/tokens/SecurityToken.sol
@@ -38,7 +38,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
 
     SemanticVersion securityTokenVersion;
 
-    // off-chain data 
+    // off-chain data
     string public tokenDetails;
 
     uint8 constant PERMISSION_KEY = 1;
@@ -58,10 +58,10 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     // Used to permanently halt all minting
     bool public mintingFrozen;
 
-    // Use to permanently halt controller actions
+    // Used to permanently halt controller actions
     bool public controllerDisabled;
 
-    // address whitelisted by issuer as controller
+    // Address whitelisted by issuer as controller
     address public controller;
 
     // Records added modules - module list should be order agnostic!
@@ -111,20 +111,31 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     event CheckpointCreated(uint256 indexed _checkpointId, uint256 _timestamp);
     // Emit when is permanently frozen by the issuer
     event FreezeMinting(uint256 _timestamp);
-    // Change the STR address in the event of a upgrade
-    event ChangeSTRAddress(address indexed _oldAddress, address indexed _newAddress);
     // Events to log minting and burning
     event Minted(address indexed _to, uint256 _value);
     event Burnt(address indexed _from, uint256 _value);
 
     // Events to log controller actions
     event SetController(address indexed _oldController, address indexed _newController);
-    event ForceTransfer(address indexed _controller, address indexed _from, address indexed _to, uint256 _value, bool _verifyTransfer, bytes _data);
-    event ForceBurn(address indexed _controller, address indexed _from, uint256 _value, bool _verifyTransfer, bytes _data);
+    event ForceTransfer(
+        address indexed _controller,
+        address indexed _from,
+        address indexed _to,
+        uint256 _value,
+        bool _verifyTransfer,
+        bytes _data
+    );
+    event ForceBurn(
+        address indexed _controller,
+        address indexed _from,
+        uint256 _value,
+        bool _verifyTransfer,
+        bytes _data
+    );
     event DisableController(uint256 _timestamp);
 
     function _isModule(address _module, uint8 _type) internal view returns (bool) {
-        require(modulesToData[_module].module == _module, "Address mismatch");
+        require(modulesToData[_module].module == _module, "Wrong address");
         require(!modulesToData[_module].isArchived, "Module archived");
         for (uint256 i = 0; i < modulesToData[_module].moduleTypes.length; i++) {
             if (modulesToData[_module].moduleTypes[i] == _type) {
@@ -222,7 +233,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
         IModuleFactory moduleFactory = IModuleFactory(_moduleFactory);
         uint8[] memory moduleTypes = moduleFactory.getTypes();
         uint256 moduleCost = moduleFactory.getSetupCost();
-        require(moduleCost <= _maxCost, "Cost too high");
+        require(moduleCost <= _maxCost, "Invalid cost");
         //Approve fee for module
         ERC20(polyToken).approve(_moduleFactory, moduleCost);
         //Creates instance of module from factory
@@ -238,9 +249,12 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
             moduleIndexes[i] = modules[moduleTypes[i]].length;
             modules[moduleTypes[i]].push(module);
         }
-        modulesToData[module] = TokenLib.ModuleData(moduleName, module, _moduleFactory, false, moduleTypes, moduleIndexes, names[moduleName].length);
+        modulesToData[module] = TokenLib.ModuleData(
+            moduleName, module, _moduleFactory, false, moduleTypes, moduleIndexes, names[moduleName].length
+        );
         names[moduleName].push(module);
         //Emit log event
+        /*solium-disable-next-line security/no-block-members*/
         emit ModuleAdded(moduleTypes, moduleName, _moduleFactory, module, moduleCost, _budget, now);
     }
 
@@ -265,8 +279,9 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     * @param _module address of module to unarchive
     */
     function removeModule(address _module) external onlyOwner {
-        require(modulesToData[_module].isArchived, "Module not archived");
+        require(modulesToData[_module].isArchived, "Not archived");
         require(modulesToData[_module].module != address(0), "Module missing");
+        /*solium-disable-next-line security/no-block-members*/
         emit ModuleRemoved(modulesToData[_module].moduleTypes, _module, now);
         // Remove from module type list
         uint8[] memory moduleTypes = modulesToData[_module].moduleTypes;
@@ -324,7 +339,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice returns a list of modules that match the provided name
+     * @notice Returns a list of modules that match the provided name
      * @param _name name of the module
      * @return address[] list of modules with this name
      */
@@ -333,7 +348,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice returns a list of modules that match the provided module type
+     * @notice Returns a list of modules that match the provided module type
      * @param _type type of the module
      * @return address[] list of modules with this type
      */
@@ -342,7 +357,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
    /**
-    * @notice allows the owner to withdraw unspent POLY stored by them on the ST or any ERC20 token.
+    * @notice Allows the owner to withdraw unspent POLY stored by them on the ST or any ERC20 token.
     * @dev Owner can transfer POLY to the ST which will be used to pay for modules that require a POLY fee.
     * @param _tokenContract Address of the ERC20Basic compliance token
     * @param _value amount of POLY to withdraw
@@ -354,6 +369,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
+
     * @notice allows owner to increase/decrease POLY approval of one of the modules
     * @param _module module address
     * @param _change change in allowance
@@ -364,7 +380,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
         uint256 currentAllowance = IERC20(polyToken).allowance(address(this), _module);
         uint256 newAllowance;
         if (_increase) {
-            require(IERC20(polyToken).increaseApproval(_module, _change), "increaseApproval fail");
+            require(IERC20(polyToken).increaseApproval(_module, _change), "IncreaseApproval fail");
             newAllowance = currentAllowance.add(_change);
         } else {
             require(IERC20(polyToken).decreaseApproval(_module, _change), "Insufficient allowance");
@@ -383,17 +399,17 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-    * @notice allows owner to change token granularity
+    * @notice Allows owner to change token granularity
     * @param _granularity granularity level of the token
     */
     function changeGranularity(uint256 _granularity) external onlyOwner {
-        require(_granularity != 0, "Incorrect granularity");
+        require(_granularity != 0, "Invalid granularity");
         emit GranularityChanged(granularity, _granularity);
         granularity = _granularity;
     }
 
     /**
-    * @notice keeps track of the number of non-zero token holders
+    * @notice Keeps track of the number of non-zero token holders
     * @param _from sender of transfer
     * @param _to receiver of transfer
     * @param _value value of transfer
@@ -403,30 +419,60 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-    * @notice removes addresses with zero balances from the investors list
-    * @param _start Index in investor list at which to start removing zero balances
-    * @param _iters Max number of iterations of the for loop
-    * NB - pruning this list will mean you may not be able to iterate over investors on-chain as of a historical checkpoint
-    */
-    function pruneInvestors(uint256 _start, uint256 _iters) external onlyOwner {
-        for (uint256 i = _start; i < Math.min256(_start.add(_iters), investorData.investors.length); i++) {
-            if ((i < investorData.investors.length) && (balanceOf(investorData.investors[i]) == 0)) {
-                TokenLib.pruneInvestors(investorData, i);
-            }
-        }
-    }
-
-    /**
      * @notice returns an array of investors
-     * NB - this length may differ from investorCount if list has not been pruned of zero balance investors
-     * @return length
+     * NB - this length may differ from investorCount as it contains all investors that ever held tokens
+     * @return list of addresses
      */
     function getInvestors() external view returns(address[]) {
         return investorData.investors;
     }
 
     /**
-     * @notice returns the investor count
+     * @notice returns an array of investors at a given checkpoint
+     * NB - this length may differ from investorCount as it contains all investors that ever held tokens
+     * @param _checkpointId Checkpoint id at which investor list is to be populated
+     * @return list of investors
+     */
+    function getInvestorsAt(uint256 _checkpointId) external view returns(address[]) {
+        uint256 count = 0;
+        uint256 i;
+        for (i = 0; i < investorData.investors.length; i++) {
+            if (balanceOfAt(investorData.investors[i], _checkpointId) > 0) {
+                count++;
+            }
+        }
+        address[] memory investors = new address[](count);
+        count = 0;
+        for (i = 0; i < investorData.investors.length; i++) {
+            if (balanceOfAt(investorData.investors[i], _checkpointId) > 0) {
+                investors[count] = investorData.investors[i];
+                count++;
+            }
+        }
+        return investors;
+    }
+
+    /**
+     * @notice generates subset of investors
+     * NB - can be used in batches if investor list is large
+     * @param _start Position of investor to start iteration from
+     * @param _end Position of investor to stop iteration at
+     * @return list of investors
+     */
+    function iterateInvestors(uint256 _start, uint256 _end) external view returns(address[]) {
+        require(_end <= investorData.investors.length, "Invalid end");
+        address[] memory investors = new address[](_end.sub(_start));
+        uint256 index = 0;
+        for (uint256 i = _start; i < _end; i++) {
+            investors[index] = investorData.investors[i];
+            index++;
+        }
+        return investors;
+    }
+
+    /**
+     * @notice Returns the investor count
+     * @return Investor count
      */
     function getInvestorCount() external view returns(uint256) {
         return investorData.investorCount;
@@ -438,15 +484,17 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     function freezeTransfers() external onlyOwner {
         require(!transfersFrozen, "Already frozen");
         transfersFrozen = true;
+        /*solium-disable-next-line security/no-block-members*/
         emit FreezeTransfers(true, now);
     }
 
     /**
-     * @notice unfreeze transfers
+     * @notice Unfreeze transfers
      */
     function unfreezeTransfers() external onlyOwner {
         require(transfersFrozen, "Not frozen");
         transfersFrozen = false;
+        /*solium-disable-next-line security/no-block-members*/
         emit FreezeTransfers(false, now);
     }
 
@@ -537,12 +585,12 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice validate transfer with TransferManager module if it exists
+     * @notice Validate transfer with TransferManager module if it exists
      * @dev TransferManager module has a key of 2
-     * @dev _isTransfer boolean flag is the deciding factor for whether the 
+     * @dev _isTransfer boolean flag is the deciding factor for whether the
      * state variables gets modified or not within the different modules. i.e isTransfer = true
      * leads to change in the modules environment otherwise _verifyTransfer() works as a read-only
-     * function (no change in the state). 
+     * function (no change in the state).
      * @param _from sender of transfer
      * @param _to receiver of transfer
      * @param _value value of transfer
@@ -584,7 +632,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice validates a transfer with a TransferManager module if it exists
+     * @notice Validates a transfer with a TransferManager module if it exists
      * @dev TransferManager module has a key of 2
      * @param _from sender of transfer
      * @param _to receiver of transfer
@@ -602,11 +650,12 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      */
     function freezeMinting() external isMintingAllowed() isEnabled("freezeMintingAllowed") onlyOwner {
         mintingFrozen = true;
+        /*solium-disable-next-line security/no-block-members*/
         emit FreezeMinting(now);
     }
 
     /**
-     * @notice mints new tokens and assigns them to the target _investor.
+     * @notice Mints new tokens and assigns them to the target _investor.
      * @dev Can only be called by the issuer or STO attached to the token
      * @param _investor Address where the minted tokens will be delivered
      * @param _value Number of tokens be minted
@@ -640,7 +689,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice mints new tokens and assigns them to the target _investor.
+     * @notice Mints new tokens and assigns them to the target _investor.
      * @dev Can only be called by the issuer or STO attached to the token.
      * @param _investors A list of addresses to whom the minted tokens will be dilivered
      * @param _values A list of number of tokens get minted and transfer to corresponding address of the investor from _investor[] list
@@ -710,7 +759,9 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     function createCheckpoint() external onlyModuleOrOwner(CHECKPOINT_KEY) returns(uint256) {
         require(currentCheckpointId < 2**256 - 1);
         currentCheckpointId = currentCheckpointId + 1;
+        /*solium-disable-next-line security/no-block-members*/
         checkpointTimes.push(now);
+        /*solium-disable-next-line security/no-block-members*/
         emit CheckpointCreated(currentCheckpointId, now);
         return currentCheckpointId;
     }
@@ -748,24 +799,25 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
      * @param _controller address of the controller
      */
     function setController(address _controller) public onlyOwner {
-        require(!controllerDisabled,"Controller disabled");
+        require(!controllerDisabled);
         emit SetController(controller, _controller);
         controller = _controller;
     }
 
     /**
-     * @notice Use by the issuer to permanently disable controller functionality
+     * @notice Used by the issuer to permanently disable controller functionality
      * @dev enabled via feature switch "disableControllerAllowed"
      */
     function disableController() external isEnabled("disableControllerAllowed") onlyOwner {
-        require(!controllerDisabled,"Controller disabled");
+        require(!controllerDisabled);
         controllerDisabled = true;
         delete controller;
+        /*solium-disable-next-line security/no-block-members*/
         emit DisableController(now);
     }
 
     /**
-     * @notice Use by a controller to execute a forced transfer
+     * @notice Used by a controller to execute a forced transfer
      * @param _from address from which to take tokens
      * @param _to address where to send tokens
      * @param _value amount of tokens to transfer
@@ -783,7 +835,7 @@ contract SecurityToken is StandardToken, DetailedERC20, ReentrancyGuard, Registr
     }
 
     /**
-     * @notice Use by a controller to execute a forced burn
+     * @notice Used by a controller to execute a forced burn
      * @param _from address from which to take tokens
      * @param _value amount of tokens to transfer
      * @param _data data to indicate validation